Add files via upload
This scratch for cs50 Problem Set 0


This  displays a road with several cars passing by. At the beginning, it asks for your name and then shows your name next to a star in the sky.
At the start, the sound of traffic plays. Once it stops, the "Tararara Trumpet" sound plays, and the text "CS50" appears next to the star.


https://scratch.mit.edu/projects/1136843563